Refactor: Extract key simulation from EvdevKeyboardMonitor (ISP)
Fixes #40

This change applies the Interface Segregation Principle by separating
key simulation from keyboard monitoring:

- Created IKeySimulator interface for key simulation operations
- Created UinputKeySimulator class with Python-based uinput simulation
- Removed SimulateKeyPressAsync, SimulateKeyComboAsync from IKeyboardMonitor
- Updated IButtonAction classes to use IKeySimulator
- Updated BluetoothMouseMonitor and UsbMouseMonitor constructors
- Registered IKeySimulator in Service DI container

Benefits:
- EvdevKeyboardMonitor reduced from 757 to 396 lines
- Clear separation of concerns (monitoring vs simulation)
- Better testability (can mock IKeySimulator independently)
- Follows ISP - clients depend only on interfaces they use
